Problème dislike non résolu

+++ backend/controllers/like.js
@@ -1,5 +1,6 @@
 //Importer models
 const db = require("../models");
+const jwt = require("jsonwebtoken");
 
 //Ajouter un like
 exports.createLike = (req, res, next) => {
@@ -25,12 +26,19 @@ exports.getAllLikesPost = (req, res, next) => {
 
 //Supprimer un like
 exports.deleteLike = (req, res, next) => {
-    //Récupérer le like de la base de données
-    db.Like.findOne({ where: { postId: req.params.postId } })
+/*  const token = req.headers.authorization.split(' ')[1];
+    const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET'); 
+    const userId = decodedToken.userId; */
+    //Récupérer le likes de la base de données
+    db.Post.findOne({ where: { id: req.params.postId} })
         .then(() => {
-            db.Like.destroy({ where: { postId: req.params.postId} })
-                .then(() => res.status(200).json({ message: 'Like supprimé !' }))
-                .catch(error => res.status(400).json({ error }));
+            db.Like.destroy({ where: { postId: req.params.postId, /* userId:userId */}})
+            /*  .then(like => {
+                console.log(like-1);
+                res.status(200).json(like-1);
+             }) */
+               .then(() => res.status(200).json({ message: 'Like supprimé !' }))
+               .catch(error => res.status(400).json({ error }));
 
         })
         .catch(error => res.status(500).json({ error }));//erreur serveur
